backend integration with frontend, ability to generate iframe
App is able to generate playlists given the client ID. Right now just generates the first playlist from there.
Included parsing for the playlist so that it gets the unique playlist ID, which is necessary for the embeded playlist iframe display (can edit the width and height, but for now it is how it is).
Need to implement putting the iframe on the second page instead of the first page (should just be a matter of refactoring).
Also there is some bug that only allows you to render the playlist array once? Or twice? but this is probably unnecessary since we are going to be generating playlists, not looking at our own.